refactor: note
