+++ src/App.tsx
@@ -3,7 +3,7 @@ import { StylesProvider, ThemeProvider } from '@material-ui/styles'
 import React, { FunctionComponent } from 'react'
 import { Route, Switch } from 'react-router'
 import { BrowserRouter } from 'react-router-dom'
-import RouteHome from './components/RouteHome'
+import RouteNote from './note/RouteNote'
 import { createTheme } from './shared/helpers/createTheme'
 
 const App: FunctionComponent = () => {
@@ -13,8 +13,8 @@ const App: FunctionComponent = () => {
         <CssBaseline />
         <BrowserRouter>
           <Switch>
-            <Route component={RouteHome} exact path={'/'} />
-            <Route component={RouteHome} path={'/:noteId'} />
+            <Route component={RouteNote} exact path={'/'} />
+            <Route component={RouteNote} path={'/:noteId'} />
           </Switch>
         </BrowserRouter>
       </ThemeProvider>
