+++ src/note/components/InputBaseNoteText.tsx
