+++ src/serviceWorker.ts
@@ -129,7 +129,7 @@ function checkValidServiceWorker(swUrl: string, config?: Config) {
     })
     .catch(() => {
       console.log(
-        'No internet connection found. RouteHome is running in offline mode.'
+        'No internet connection found. RouteNote is running in offline mode.'
       )
     })
 }
